+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.1.1
 
 - `flush` operation to move all or a subset of in-memory blobs to disk.
+- Implemented a `blobId` cache to reduce disk access.
 - Consistency fix: `BlobStorage.getSize()` returns `null` instead of `-1` when blob doesn't exist. (**Breaking change.**)
 
 ## 0.1.0